requirements.txt
jinja2
requests
celery
pyyaml
boto3
